Add files via upload
The concept of smoke detector integrated for fire emergency hazard systems is quite well known 
in the current era. Due to the serious health, safety, environmental and economic issues of the 
release of hazardous materials, the device that controls sensors, processes information, produces 
evacuation alarms and shuts down equipment and gas valves can be easy to use. Besides, most of 
death cause by fire incident is not by fire burn but by excessive smoke inhalation. In this work
there has been the development of a low cost, and reliable microcontroller based automated fire 
alarm system for remotely alerting any fire incidents in a potential fire arising areas, Engineering 
laboratory (UENR) as the area of study. A review of an existing fire-detector type (MQ-2 sensor 
module) has been carried out along with a compacted microcontroller board (Arduino Nano) 
based on the ATmega328P microcontroller, buzzer, a sim module (sim 800L), a buck convertor, 
and the development of changeover system. The main aim of the system is to alert the distant 
user quickly by sending a message (SMS) via GSM network. The Arduino programming 
language is use, which is a simplified version of C++. The code is written in the Arduino 
Integrated Development Environment (IDE) and then converted to the C language and uploaded 
to the Arduino Nano board and initiated the codes to the various modules on the device which 
allows sensor to detect smoke, when buzzer should beep, and when sim module should also send 
notification to the user that there is a smoke detected. This system would help all the users at any 
level of income to have one at their home and it save great losses and damage which might 
happen due to fire.